Solved likes!

+++ webui/src/components/PostCard.vue
@@ -12,40 +12,41 @@ export default {
 			likersResult: [],
 			commentsResults: [],
 			likesBase: "",
-			commentsLimit: ""
+			commentsLimit: "",
+			username: localStorage.username
         };
     },
     methods: {
 		async refreshData(new_likes=false, new_comments=false) {
+			if (!this.imgId) {
+				return
+			}
 			this.errormsg = null;
 			try {
 				var response = await this.$axios.get("/photos/" + this.imgId.toString(), {params: {"likes-base": this.likesBase, "comments-limit":this.commentsLimit}});
 				this.title = response.data["photo-title"]
 				this.author = response.data["photo-author"]
-				var auLink = document.getElementById("author-link")
-				auLink.onclick = this.moveToAuthor;
 				this.date = this.convDate(response.data["photo-date"])
 				this.liked = response.data["liked"]
-				var but = document.getElementById("button-like")
-				if (this.liked) {
-					but.textContent = "unlike"
-					but.classList.add("post-unlike-button")
-					but.classList.remove("post-like-button")
-					but.onclick = () => this.unlikeThis().then()
-				} else {
-					but.textContent = "like"
-					but.classList.remove("post-unlike-button")
-					but.classList.add("post-like-button")
-					but.onclick = () => this.likeThis().then()
-				}
 				
+				for (let i=0; i< response.data["likes-running-batch"].length; i++) {
+					if (response.data["likes-running-batch"][i].name == this.username) {
+						response.data["likes-running-batch"].splice(i, 1)
+						break
+					}
+				}
+
 				if (new_likes) {
 					this.likersResult.push(...response.data["likes-running-batch"])
-					this.likesBase = response.data["likes-running-batch"][response.data["likes-running-batch"].length - 1]["name"];
+					if (response.data["likes-running-batch"].length > 0) {
+						this.likesBase = response.data["likes-running-batch"][response.data["likes-running-batch"].length - 1].name;
+					}
 				}
 				if (new_comments) {
 					this.commentsResults.push(...response.data["comments-running-batch"])
-					this.commentsLimit = response.data["comments-running-batch"][response.data["comments-running-batch"].length - 1]["comment-id"];
+					if (response.data["comments-running-batch"].length > 0) {
+						this.commentsLimit = response.data["comments-running-batch"][response.data["comments-running-batch"].length - 1]["comment-id"];
+					}
 				}
 
 			} catch (e) {
@@ -53,36 +54,30 @@ export default {
 			}
 		},
 		async likeThis() {
+			if (!this.imgId) {
+				return
+			}
 			try {
 				await this.$axios.put("/photos/" + this.imgId.toString() + "/likes/" + localStorage.username);
 			} catch (e) {
 				this.errormsg = e.toString();
 			}
 			this.refreshData();
-			this.disableLike();
 		},
 		async unlikeThis() {
+			if (!this.imgId) {
+				return
+			}
 			try {
 				await this.$axios.delete("/photos/" + this.imgId.toString() + "/likes/" + localStorage.username);
 			} catch (e) {
 				this.errormsg = e.toString();
 			}
 			this.refreshData();
-			this.disableLike();
-		},
-		disableLike() {
-			var but = document.getElementById("button-like")
-			but.disabled = true;
-			setTimeout(function() {
-				but.disabled = false;
-			}, 500);
 		},
 		convDate(date) {
 			return date //TODO: convert to local timezone
 		},
-		moveToAuthor() {
-			this.$router.push({name: 'user', params: {username: this.author}})
-		}
     },
 	created() {
 		if (this.imgId) {
@@ -103,9 +98,7 @@ export default {
 				</div>
 				<div class="post-info">
 					<div>					
-						<div id="author-link" class="post-author" style="cursor: pointer;" align="right">
-							{{ author }}
-						</div>
+						<router-link class="post-author" align="right" :to="'/users/' + author">{{ author }}</router-link>
 						<div class="post-date" align="right">
 							{{ date }}
 						</div>
@@ -117,10 +110,11 @@ export default {
 		<div class="post-commentsandlikes">
 			<div id="likesholder" class="post-likesholder">
 				<router-link class="like-element" v-for="element in likersResult" :to="'/users/' +element['name']">{{ element['name'] }}</router-link>
+				<router-link v-if="liked" class="like-element" :to="'/users/' + username">{{ username }}</router-link>
+				<button class="like-element-end" :onclick="() => this.refreshData(new_likes=true)">More likers!</button>
 			</div>
-			<button id="button-like" class="post-like-button">
-				like
-			</button> 
+			<button v-if="!liked" class="post-like-button" :onclick="() => this.likeThis().then()">like</button>
+			<button v-if="liked" class="post-unlike-button" :onclick="() => this.unlikeThis().then()">unlike</button>
 			<div class="post-comments">
 			</div>
 			<textarea class="post-comment-writer" maxlength="256"></textarea>
@@ -131,23 +125,3 @@ export default {
 		</div>
 	</div>
 </template>
-
-<!-- this.likersResult.forEach(element => {
-	var liker = element["name"]
-	var but = document.createElement("button");
-	but.addEventListener('click', () => this.$router.push({name: 'user', params: {username: liker}}));
-	var text = document.createTextNode(liker);
-	but.classList.add("like-element")
-	but.appendChild(text);
-	holder.appendChild(but);
-});
-
-if (this.likersResult.length >= 64) {
-	var but = document.createElement("button");
-	but.addEventListener('click', () => this.refreshData(new_likes=true));
-	but.setAttribute("id", "furtherLikes_button")
-	var text = document.createTextNode("More likers!");
-	but.classList.add("like-element-end")
-	but.appendChild(text);
-	holder.appendChild(but);
-} -->
